// ecommerce/src/data/store/interfaces/product.ts
export interface Product {
    id: number
    ean: string
    name: string
    code: string
    description: string
    categoryId: number
    mediaPath: string
    mediaName: string
    storeId: number,
    stock: string
    sellingPrice: string
    promotionPrice: string
    categoryName: string
    companyId: number
}